[WFCORE-1618] Get rid of pointless package protected FutureManagementChannel.Establishing; push code down to sole subclass
